trying to figure out graph
added arraylist of filenames to graph class to keep track of files.
changed format of number of vertices so that every time there is a new
file add a vertice. updated functions in Graph so that you can pass in a
String not an integer.  I have an idea for alphabetical order